requirements.txt
pandas>=1.1.4
Flask>=1.1.1
flask-cors
gunicorn>=19.9.0
Flask-HTTPAuth>=4.1.0
fastparquet
google-cloud-bigquery
